// project/Code_v2/src/types.ts
import type { Hypothesis, SpeechStateExternalEvent } from "speechstate";
import type { ActorRef } from "xstate";

export interface DMContext {
  spstRef: ActorRef<any, any>;
  lastResult: Hypothesis[] | null;
  pt_machine: number | 0;
  pt_user: number | 0;
  name: string | null;
  current_motivation: string | null;
  // interpretation: NLUObject | null;
  current_word: string | null;
  association_found: boolean | null;
  previous_words: Array<string> | string[];
}

export interface Entity { // This is the type of the entities array in the NLUObject.
  category: string;
  text: string;
  confidenceScore: number;
  offset: number;
  length: number;
}

export interface Intent { // This is the type of the intents array in the NLUObject.
  category: string;
  confidenceScore: number;
}

export interface NLUObject { // This is the type of the interpretation in the DMContext.
  entities: Entity[];
  intents: Intent[];
  projectKind: string;
  topIntent: string;
}

export type DMEvents = SpeechStateExternalEvent | { type: "CLICK" } | { type: "DONE" };